refactor: auth  module created
