+++ src/app/app.module.ts
@@ -11,6 +11,7 @@ import { AppComponent } from './app.component';
 import { AppInitService } from './core/service/app-init.service';
 import { BaseModule } from './base/base.module';
 import { SettingsModule } from './settings/settings.module';
+import { AuthModule } from './auth/auth.module';
 import { AuthInterceptor } from './core/interceptors/auth.interceptor';
 import { ErrorComponent } from './shared/components/error/error.component';
 import { WarningComponent } from './shared/components/warning/warning.component';
@@ -34,6 +35,7 @@ export function initializeApp1(appInitService: AppInitService) {
     HttpClientModule,
     BaseModule,
     SettingsModule,
+    AuthModule,
     TranslateModule.forRoot({
       loader: {
         provide: TranslateLoader,
