+++ src/app/auth/auth.module.ts
@@ -1,0 +1,15 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { SharedModule } from '../shared/shared.module';
+import { AuthPageComponent } from './pages/auth-page/auth-page.component';
+
+@NgModule({
+  declarations: [AuthPageComponent],
+  imports: [
+    CommonModule,
+    SharedModule
+  ],
+  exports: [AuthPageComponent]
+})
+export class AuthModule { }
