+++ src/app/auth/pages/auth-page/auth-page.component.scss
