+++ src/app/base/pages/content-page/content-page.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { AuthPageComponent } from '../auth-page/auth-page.component';
+import { AuthPageComponent } from '../../../auth/pages/auth-page/auth-page.component';
 import { UserProfileComponent } from '../../../settings/pages/user-profile/user-profile.component';
 import { ThemesListComponent } from '../../components/themes-list/themes-list.component';
